requirements.txt
django-cors-headers==3.12.0
django-heroku==0.3.1
djangorestframework==3.13.1
email-validator==1.2.1
gunicorn==20.1.0
pillow==8.3.1
pyjwt==2.1.0
pyuploadcare==3.0.0
